requirements.txt
Django>=3.0.0,<3.1.0
djangorestframework>=3.11.0, <3.12.0
django-cors-headers>=3.4.0,<3.5.0
